+++ src/search/mode.rs
@@ -17,3 +17,33 @@ pub trait Mode: core::fmt::Debug {
 pub struct Enumerate;
 
 impl Mode for Enumerate {}
+
+/// Enumerate assignments that satisfy all constraints, and gradually lower objective expression.
+#[derive(Debug)]
+pub struct Minimize<V> {
+    objective: V,
+    minimum_opt: Option<i32>,
+}
+
+impl<V: View> Minimize<V> {
+    pub const fn new(objective: V) -> Self {
+        Self {
+            objective,
+            minimum_opt: None,
+        }
+    }
+}
+
+impl<V: View> Mode for Minimize<V> {
+    gen fn on_branch(&self, space: &mut Space) -> PropId {
+        // Prune assignments that cannot lower objective expression
+        if let Some(minimum) = self.minimum_opt {
+            yield space.props.less_than(self.objective, minimum);
+        };
+    }
+
+    fn on_solution(&mut self, vars: &Vars) {
+        // New objective value is necessarily lower than previous lowest
+        self.minimum_opt = Some(self.objective.min_raw(vars));
+    }
+}