+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Emoji prefixes follow the [gitmoji](https://gitmoji.dev/) guide.
 ### Added
 
 - ✨ Implement base propagation engine
+- ✨ Offer different modes: solve, enumerate, minimize (or maximize) with intermediate assignments
 - 🚸 Make propagators more generic using views
 - 🔨 Harden [clippy](https://github.com/rust-lang/rust-clippy) configuration
 - 🔨 Declare common commands as [just](https://github.com/casey/just) recipes